TST: Import sqlalchemy on Travis.
