+++ ci/requirements-2.6.txt
@@ -4,3 +4,4 @@ python-dateutil==2.1
 pytz==2013b
 http://www.crummy.com/software/BeautifulSoup/bs4/download/4.2/beautifulsoup4-4.2.0.tar.gz
 html5lib==1.0b2
+sqlalchemy==0.8
