+++ ci/requirements-2.7.txt
@@ -16,3 +16,4 @@ scikits.timeseries==0.91.3
 MySQL-python==1.2.4
 scipy==0.10.0
 beautifulsoup4==4.2.1
+sqlalchemy==0.8
